Modify to get correct correspondence of IRs for rotations on arms
Before IRs for arms are obtained independently, but for some pont groups
it was difficult to get correct correspondence.
For example, for the mm2 point group, sgv and sgv' generally give different
IRs, and correct correspondence of symmetry operations must be considered
between different arms in the star.
Now we first fix the representative of the star, and classes of rotational
operations are considered based on the representative.