Fix to calculate average masses for each inequivalent site
Before symmetrically inequivalent sites for "unitcell_ideal" were not
distinguished; the averages are done only according to the number of atoms for
each dummy "element".
This might cause problems for systems with two or more sublattices such
as the omega structure and the the dhcp structure.
Now the averages are taken over each set of symmetrically equivalent sites.