Modify not to raise Warning
Before in `_search_primitive_symmetry_ideal` we raised a Warning when
the numbers of symmetry operations are different between the supercell
and the primitive cell. This guaranteed that the unfolded band structure
has the same symmetry as the primitive cell. This could be however too
strong restriction for some purpose, and therefore I now test to remove
this restriction.